requirements.txt
streamlit
pandas
python-dotenv
requests
openpyxl
plotly